pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.6.2",
    "celery-types>=0.24.0",
    "channels>=4.3.2",
    "channels-redis>=4.3.0",
    "cloudinary>=1.44.1",
    "decouple>=0.0.7",
    "django>=6.0.2",
    "django-autoslug>=1.9.9",
    "django-celery-beat>=2.9.0",
    "django-cloudinary-storage>=0.3.0",
    "django-cors-headers>=4.9.0",
    "django-countries>=8.2.0",
    "django-database-url>=1.0.3",
    "django-debug-toolbar>=6.2.0",
    "django-environ>=0.13.0",
    "django-filter>=25.2",
    "django-phonenumber-field>=8.4.0",
    "django-redis>=6.0.0",
    "django-silk>=5.4.3",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "djangorestframework-stubs>=3.16.8",
    "dotenv>=0.9.9",
    "drf-nested-routers>=0.95.0",
    "drf-spectacular>=0.29.0",
    "drf-yasg>=1.21.15",
    "email-validator>=2.3.0",
    "factory-boy>=3.3.3",
    "faker>=40.5.1",
    "gunicorn>=25.1.0",
    "phonenumberslite>=9.0.25",
    "postal-regex>=1.0.0",
    "psycopg2-binary>=2.9.11",
    "pytest>=9.0.2",
    "pytest-django>=4.12.0",
    "requests>=2.32.5",
    "sendgrid>=6.12.5",
    "uvicorn[standard]>=0.41.0",
    "validators>=0.35.0",
    "whitenoise>=6.12.0",
]
